# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeavesFlow API",
    "description": "LeavesFlow MVP V1 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me": {
      "get": {
        "summary": "Get Me",
        "operationId": "get_me_api_v1_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Me",
        "operationId": "update_me_api_v1_me_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/tag-profile": {
      "put": {
        "summary": "Put Tag Profile",
        "operationId": "put_tag_profile_api_v1_me_tag_profile_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTagProfileIds"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTagProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals": {
      "post": {
        "summary": "Create Goal",
        "operationId": "create_goal_api_v1_goals_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}/plan:generate": {
      "post": {
        "summary": "Generate Plan",
        "operationId": "generate_plan_api_v1_goals__goal_id__plan_generate_post",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}": {
      "get": {
        "summary": "Get Goal",
        "operationId": "get_goal_api_v1_goals__goal_id__get",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}/plan": {
      "get": {
        "summary": "Get Plan",
        "operationId": "get_plan_api_v1_goals__goal_id__plan_get",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/active-plan": {
      "get": {
        "summary": "Get Active Plan",
        "operationId": "get_active_plan_api_v1_me_active_plan_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivePlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/check-ins": {
      "post": {
        "summary": "Create Check In",
        "operationId": "create_check_in_api_v1_tasks__task_id__check_ins_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "tagProfile": {
            "$ref": "#/components/schemas/UserTagProfileOut"
          }
        },
        "type": "object",
        "required": [
          "token",
          "user",
          "tagProfile"
        ],
        "title": "AuthResponse"
      },
...
      "CreateGoalRequest": {
        "properties": {
          "rawInput": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Rawinput"
          },
          "profileSnapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserTagProfileIds"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "rawInput"
        ],
        "title": "CreateGoalRequest"
      },
      "CreateGoalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rawInput": {
            "type": "string",
            "title": "Rawinput"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed"
            ],
            "title": "Status"
          },
          "profileSnapshot": {
            "$ref": "#/components/schemas/UserTagProfileOut"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "rawInput",
          "status",
          "profileSnapshot",
          "createdAt"
        ],
        "title": "CreateGoalResponse"
      },
      "GeneratePlanRequest": {
        "properties": {},
        "type": "object",
        "title": "GeneratePlanRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MeResponse": {
        "properties": {
          "userId": {
            "type": "string",
            "title": "Userid"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "tagProfile": {
            "$ref": "#/components/schemas/UserTagProfileOut"
          },
          "skillTags": {
            "items": {
              "$ref": "#/components/schemas/SkillTagOut"
            },
            "type": "array",
            "title": "Skilltags"
          },
          "activePlan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivePlanResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "userId",
          "user",
          "tagProfile",
          "skillTags"
        ],
        "title": "MeResponse"
      },
...
      "RegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "profile": {
            "$ref": "#/components/schemas/UserTagProfileIds"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "profile"
        ],
        "title": "RegisterRequest"
      },
...
      "UpdateMeRequest": {
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Displayname"
          }
        },
        "type": "object",
        "required": [
          "displayName"
        ],
        "title": "UpdateMeRequest"
      },
      "UpdateTagProfileResponse": {
        "properties": {
          "tagProfile": {
            "$ref": "#/components/schemas/UserTagProfileOut"
          }
        },
        "type": "object",
        "required": [
          "tagProfile"
        ],
        "title": "UpdateTagProfileResponse"
      },
...
      "UserTagProfileIds": {
        "properties": {
          "identityTagIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Identitytagids"
          },
          "backgroundTagIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Backgroundtagids"
          },
          "levelTagIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Leveltagids"
          },
          "goalTypeTagIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Goaltypetagids"
          },
          "timeRangeTagIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Timerangetagids"
          },
          "outputPreferenceTagIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Outputpreferencetagids"
          }
        },
        "type": "object",
        "title": "UserTagProfileIds"
      },
...
    }
  }
}